Create custom EEPROM emulator
TI's FEE driver failed for some reason and I wasn't able to explain it. Every write was failing until a system reset happened where the data would update but with the wrong FEE checksum
We don't need the complexity of the FEE driver since we are storing very little data, so I created a primitive EEPROM emulator using the F021 Flash API. It has 32 sectors each 4kb in size.
I also ported all of the code that accesses the eeprom. I changed the interface to work with pointers for performance reasons and to allow error code returns